Renaming: tng-sdk-profile to tng-sdk-benchmark
