+++ README.md
@@ -2,9 +2,9 @@
 
 <p align="center"><img src="https://github.com/sonata-nfv/tng-api-gtw/wiki/images/sonata-5gtango-logo-500px.png" /></p>
 
-# tng-sdk-profile
+# tng-sdk-benchmark
 
-This repository contains the `tng-sdk-profile` component that is part of the European H2020 project [5GTANGO](http://www.5gtango.eu) NFV SDK. This component is responsible to automatically execute performance benchmarks of NFV network services and functions.
+This repository contains the `tng-sdk-benchmark` component that is part of the European H2020 project [5GTANGO](http://www.5gtango.eu) NFV SDK. This component is responsible to automatically execute performance benchmarks of NFV network services and functions.
 
 The seed code of this component is based on the `son-cli` toolbox, specifically the `son-profile` tool, which was developed as part of the European H2020 project [SONATA](http://sonata-nfv.eu).
