+++ pipeline/promote/promote-int.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -e
-docker tag registry.sonata-nfv.eu:5000/tng-sdk-profile:latest registry.sonata-nfv.eu:5000/tng-sdk-profile:int
-docker push registry.sonata-nfv.eu:5000/tng-sdk-profile:int
+docker tag registry.sonata-nfv.eu:5000/tng-sdk-benchmark:latest registry.sonata-nfv.eu:5000/tng-sdk-benchmark:int
+docker push registry.sonata-nfv.eu:5000/tng-sdk-benchmark:int
