+++ pipeline/unittest/test.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 set -e
-docker run -i --rm registry.sonata-nfv.eu:5000/tng-sdk-profile pytest -v
+docker run -i --rm registry.sonata-nfv.eu:5000/tng-sdk-benchmark pytest -v
