+++ src/tngsdk/benchmark/generator/template/sonata_nsd_function_experiment.yml
