+++ src/tngsdk/benchmark/generator/template/sonata_vnfd_mp.yml
